Improve expected size of c++ data, and add explicit casts.
This eliminates most compiler warnings.

+++ CorsixTH/Src/sdl_core.cpp
@@ -92,9 +92,9 @@ struct fps_ctrl
     {
         Uint32 now = SDL_GetTicks();
         frame_time[q_front] = now;
-        q_front = (q_front + 1) % (sizeof(frame_time) / sizeof(*frame_time));
+        q_front = (q_front + 1) % static_cast<int>((sizeof(frame_time) / sizeof(*frame_time)));
         if(q_front == q_back)
-            q_back = (q_back + 1) % (sizeof(frame_time) / sizeof(*frame_time));
+            q_back = (q_back + 1) % static_cast<int>((sizeof(frame_time) / sizeof(*frame_time)));
         else
             ++frame_count;
         if(now < 1000)
@@ -104,7 +104,7 @@ struct fps_ctrl
         while(frame_time[q_back] < now)
         {
             --frame_count;
-            q_back = (q_back + 1) % (sizeof(frame_time) / sizeof(*frame_time));
+            q_back = (q_back + 1) % static_cast<int>((sizeof(frame_time) / sizeof(*frame_time)));
         }
     }
 };
